pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "gtts>=2.5.4",
    "joblib>=1.5.2",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pygame>=2.6.1",
    "pyttsx3>=2.99",
    "speechrecognition>=3.14.3",
    "streamlit>=1.50.0",
    "tensorflow>=2.18.1",
]
